requirements.txt
openai
python-dotenv
chromadb
